Move everything related to the Flash middleware:
- Since the Flash Middleware is no longer a middleware (see
  https://github.com/rails/rails/pull/44588#issuecomment-1055845413
  for the reason) it doesn't make sense to have it in the Middleware
  folder.

  Also decided to split all nested classes inside separate files,
  make things easier to read.

  This commit is in prepration to deprecated the `Flash::Middleware`
  constant.
